+++ tests/meltano/core/test_elt_context.py
@@ -46,6 +46,7 @@ def assert_transformer_env(transformer, env):
     assert (
         env["MELTANO_TRANSFORM_TARGET"]
         == env["DBT_TARGET"]
+        == env["MELTANO_LOAD__DIALECT"]
         == env["MELTANO_LOADER_NAMESPACE"]
     )
     assert (